# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Aurion OS API
  description: |
    # Aurion OS API Documentation
    
    Complete API for the Aurion Operating System including:
    - 🤖 JARVIS AI Assistant with emotional intelligence and autonomy
    - 🔐 Authentication and User Management
    - 🛡️ VPN Services
    - 💰 Financial Operations
    - 🧠 Cognitive Engine
    - 🌐 OSINT Tools
    
    ## Base URL
    Production: `https://api.aurionai.ru/v1`
    
    ## Authentication
    All endpoints require Bearer token authentication:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
  version: 1.0.0
  contact:
    name: Aurion OS Support
    email: support@aurionai.ru
    url: https://www.aurionai.ru
servers:
  - url: https://api.aurionai.ru/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register new user
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        400:
          description: Invalid input
        409:
          description: User already exists
  /auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: Invalid credentials
  /auth/refresh:
    post:
      summary: Refresh access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        200:
          description: Token refreshed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - username
        - password
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 8
          format: password
        invite_code:
          type: string
          description: Optional invite code
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        totp_code:
          type: string
          description: 2FA code if enabled
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
          description: Seconds until expiration
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [user, creator, admin]
        subscription:
          type: string
          enum: [free, basic, premium, enterprise]
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time

    # ===== JARVIS SCHEMAS =====
...